descent: port FlowType, SymbolType, Transaction, Union/Intersection/DifferenceAddressSetView
Java classes ported:
- ghidra.program.model.symbol.FlowType: newtype over the already-ported RefType enum
  (RefType already unifies FlowType's and DataRefType's constant values), exposing
  FlowType's own query-method overrides.
- ghidra.program.model.symbol.SymbolType: replaced the old label-only enum with a full
  port including isValidParent/isValidAddress/isValidSourceType per-instance behavior,
  allowsDuplicates, isNamespace, and getSymbolType(id) lookup. Required growing
  Symbol::get_program() (defaulted to None) and adding Address::is_variable_address().
  SymbolType::get_id() is now i8 (was i32) to match Java's byte getID(), so a few call
  sites that already did `as i8` were untouched and the two that needed i32 were fixed
  (simple_diff_utility.rs, symbol_database_adapter.rs); from_id/display_name callers
  were renamed to get_symbol_type/name to match the real Java method names.
- db.Transaction: RAII guard (Transaction<T: EndTransaction>) mirroring LockHold's
  existing precedent for AutoCloseable ports; Drop stands in for try-with-resources.
- ghidra.util.{Union,Intersection,Difference}AddressSetView: lazily-recomputed
  AddressSetView combinators layered on AbstractAddressSetView. The real lazy merge
  machinery (AddressRangeIterators' TwoWayBreakdownAddressRangeIterator/
  UnionAddressRangeIterator) is still TODO in PORT_MANIFEST, so per the precedent
  already documented in abstract_address_set_view.rs, the three range-producing
  methods recompute the same result eagerly via the already-tested AddressSet
  union/intersect/subtract operations; every method each Java subclass explicitly
  overrides (contains variants, isEmpty/min/max for Union, getRangeContaining for
  Intersection/Difference) is reproduced faithfully, everything else inherits
  AbstractAddressSetView's defaults exactly as in Java.

Also confirmed ThunkReference and MemReferenceImpl were already fully and faithfully
ported in a prior session but never flipped to DONE in the manifest; verified against
the Java sources and flipped now.

Quirk faithfully reproduced: SymbolType.getSymbolType(2) returns null in Java because
id 2 is a deliberate hole in the backing constant array (never assigned to a type);
SymbolType::get_symbol_type(2) returns None to match, with a dedicated test.

+++ ghidra-rs/src/program/util/simple_diff_utility.rs
@@ -792,8 +792,8 @@ fn compare_matches(a: &ExternalReferenceCount, b: &ExternalReferenceCount) -> st
     b.rank
         .cmp(&a.rank)
         .then_with(|| {
-            let a_id = a.symbol_type().map(|t| t.get_id()).unwrap_or(i32::MIN);
-            let b_id = b.symbol_type().map(|t| t.get_id()).unwrap_or(i32::MIN);
+            let a_id = a.symbol_type().map(|t| t.get_id()).unwrap_or(i8::MIN);
+            let b_id = b.symbol_type().map(|t| t.get_id()).unwrap_or(i8::MIN);
             b_id.cmp(&a_id)
         })
         .then_with(|| {
